@@ -48,6 +48,7 @@ test {
 	useJUnit {
 		dependencies {
 			compile(group: 'junit', name: 'junit', version: '[4,)')
+			testcompile group: 'org.xerial', name: 'sqlite-jdbc', version: '[3,)'
 		}
 	}
 }
@@ -59,17 +60,17 @@ buildscript {
 		}
 	}
 	dependencies {
-		classpath(group: 'org.jfrog.buildinfo', name: 'build-info-extractor-gradle', version: '[2,)')
+//classpath(group: 'org.jfrog.buildinfo', name: 'build-info-extractor-gradle', version: '[2,)')
 		classpath "org.kt3k.gradle.plugin:coveralls-gradle-plugin:[2,)"
-		classpath 'com.moowork.gradle:gradle-node-plugin:0.9'
+		classpath 'com.moowork.gradle:gradle-node-plugin:[1,)'
 	}
 }
-dependencies {
-	compile group: 'org.webjars', name: 'jquery', version: '2.1.4'
-}
+//dependencies {
+//	compile group: 'org.webjars', name: 'jquery', version: '2.1.4'
+//}
 
 artifacts { 
-	archives buildCoreJar16, buildCoreJar17, buildCoreJar18, buildSourceJar, buildMinCoreJar, buildJavadoc
+	archives buildCoreJar16, buildCoreJar17, buildCoreJar18, buildCoreJar19, buildSourceJar, buildMinCoreJar, buildJavadoc
 }
 
 test {
@@ -82,7 +83,7 @@ test {
 	}
 }
 task wrapper(type: Wrapper) {
-    gradleVersion = '4.1'
+		gradleVersion = '4.4'
 }
 jacoco.toolVersion = "0.7.+"
 
@@ -121,8 +122,8 @@ if (project.hasProperty('branch')) {
 test.finalizedBy jacocoTestReport
 
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 1.9
+targetCompatibility = 1.9
 
 defaultTasks 'test', 'jacocoTestReport', 'buildJavadoc'
 node {
