@@ -30,20 +30,42 @@ public class ModelHistory {
 		return history.last();
 	}
 
-	// TODO Functionality
-	public void refactoringHistory() {
-		boolean refactoring = true;
-		int lowestId = 0;
+	public boolean refactoringHistory() {
+//		int lowestId = 0;
 		SortedSet<NodeProxy> nodes = getSpace().getNodeProxies();
-		for (NodeProxy proxy : nodes) {
-			if (!proxy.isOnline()) {
-				refactoring = false;
-				break;
+		SimpleKeyValueList<String, Integer> keys = new SimpleKeyValueList<String, Integer>();
+		for(ModelChange change : history) {
+			keys.add(change.getKey(), 0);
+			if(change.getChange() instanceof JsonObject == false) { 
+				return false;
 			}
 		}
+		for (NodeProxy proxy : nodes) {
+//			if (!proxy.isOnline()) {
+//				return false;
+//			}
+			String key = proxy.getHistory();
+			int pos = keys.indexOf(key);
+			if(pos >= 0) {
+				keys.setValue(pos, keys.get(key) + 1);
+			}
+		}
+		// Now refacotring
+//		int refactoring
+		//TODO REFACOTRING
+		for(ModelChange change : history) {
+			int start=0;
+			int pos = 0;
+			while(keys.getValueByIndex(pos) == 0) {
+				BaseItem change2 = change.getChange();
+			}
+//			change.
+//			keys.add(change.getKey(), 0);
+		}
+		return true;
 	}
 
-	private boolean isToManyField(SendableEntityCreator createrClass, String fieldName) {
+	protected boolean isToManyField(SendableEntityCreator createrClass, String fieldName) {
 		Object prototype = prototypeCache.get(createrClass);
 
 		if (prototype == null) {
